docs: api url - updated the production API url across the project, especially in the dashboard and the vscode extension - released a new version (`0.0.19`) of the extension with the new api - updated the refs to the API url in the READMEs - moved `PeriodProjects.tsx` to `dashboard-page`
