+++ apps/api/README.md
@@ -4,7 +4,7 @@
 
 <h1 align="center">MoonCode API</h1>
 <p align="center">The API for the MoonCode app<br/>
-<a href="https://mooncode-api.fly.dev">mooncode-api.fly.dev</a>
+<a href="https://api.mooncode.cc">api.mooncode.cc</a>
 </p>
 
 <p align="center">
@@ -13,36 +13,42 @@
 </p>
 
 ## Description
+
 This project is the API used to power the other main packages in the monorepo: [dashboard](../dashboard) and [vscode-extension](../vscode-extension).
 It is built on top of [Nestjs](https://nestjs.com/) and powered by [trpc](https://trpc.io/).
 
 ## Project setup
 
-To run the API, you need to first clone the repository 
-``` bash
+To run the API, you need to first clone the repository
+
+```bash
 git clone https://github.com/Friedrich482/mooncode-monorepo.git
 ```
+
 Then `cd` in the API
 
 ```bash
 cd apps/api
 ```
+
 Then install dependencies
 
 ```bash
 npm install
 ```
+
 Before continuing you'll need some environment variables: `JWT_TOKEN` and `DATABASE_URL`.
 
 ### Environment variables
+
 - `DATABASE_URL`, you can use:
 
   ```bash
   DATABASE_URL="postgresql://postgres:postgres@localhost:5432/mooncode"
   ```
-  
+
   [compose.yaml](./compose.yaml)
-  
+
   (Local database set up by docker compose as a docker volume. If you don't have docker installed, follow that link to the installation: [Docker](https://docs.docker.com/get-started/get-docker/))
 
 - `JWT_TOKEN`: you need to generate a SSH Key and get the fingerprint.
@@ -53,11 +59,12 @@ Before continuing you'll need some environment variables: `JWT_TOKEN` and `DATAB
 
 ## Compile and run the project
 
-Compile : 
+Compile :
 
 ```bash
 npm run build
 ```
+
 Then start:
 
 ```bash
@@ -66,19 +73,20 @@ npm run start:prod
 
 ## Deployment
 
-The API is currently deployed on [fly.io](https://fly.io).
+The API is currently deployed on [api.mooncode.cc](https://api.mooncode.cc).
 
-## Dockerisation
+## Containerization
 
 To dockerize the application, you need to place yourself at the root of the monorepo, then
 
 ```bash
-docker build -f apps/api/Dockerfile -t mooncode-api --progress=plain .
+docker build -t mooncode-api -f apps/api/Dockerfile --progress=plain .
 ```
-And to run a container called `mooncode-api-container`(from `apps/api`) : 
+
+And to run a container called `mooncode-api-container` :
 
 ```bash
-docker run -p 3000:3000 --name mooncode-api-container --env-file .env mooncode-api
+docker run -p 3000:3000 --name mooncode-api-container --env-file apps/api/.env mooncode-api:latest
 ```
 
 ## License
